err9

+++ vdhost/cli/stop_miner.py
@@ -2,6 +2,7 @@ import click
 import subprocess
 import os
 import json
+import xml.etree.ElementTree as ET
 
 from colored import fg
 from colored import stylize
@@ -58,15 +59,34 @@ def stop_miner(gpu_id):
             #args = ['kill', '--', '-$(ps', '-o', 'pgid=', str(pid), '|', 'grep', '-o', '[0-9]*)']
             #args = ['kill', '--', '-'+str(pid)
             #subprocess.check_call(args)
-            pgrp = os.getpgid(pid)
-            os.killpg(pgrp, signal.SIGTERM)
+            
+            #pgrp = os.getpgid(pid)
+            #os.killpg(pgrp, signal.SIGTERM)
+            result = subprocess.run(['nvidia-smi', '-q', '-i', str(gpu_id), '-x'], stdout=subprocess.PIPE)
+            result = result.stdout.decode()
 
-            # If the pids have not yet been killed, try again
-            while pid_exists(str(pid)):
-                print("Attempting to stop mining")
-                #args2 = ['kill', '-9', '-p', str(pid)]
-                #subprocess.check_call(args2)
-                os.killpg(pgrp, signal.SIGKILL)
+            # no matching GPUs, so we don't need to stop any processes
+            if "No devices" in result:
+                return
+
+            # parsing the XML tree response
+            root = ET.fromstring(result)
+            gpu = root.find('gpu')
+            processes = gpu.find('processes')
+
+            # killing each of the processes
+            for process in processes:
+                pid = int(process.find('pid').text)
+                pgrp = os.getpgid(pid)
+
+                os.killpg(pgrp, signal.SIGTERM)
+
+                # If the pids have not yet been killed, try again
+                while pid_exists(str(pid)):
+                    print("Attempting to stop mining")
+                    #args2 = ['kill', '-9', '-p', str(pid)]
+                    #subprocess.check_call(args2)
+                    os.killpg(pgrp, signal.SIGKILL)
 
             # update dict
             #pid_dat[gpu_id] = -1
